requirements.txt
# Django
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Requests
requests==2.32.3

# Photo & Image
pillow==11.0.0